refactor: redirect to component
